add remaining flashpro sensors and sensor metadata

+++ Datalog.OBDII/IOBDIIDatalogFrame.cs
@@ -10,7 +10,7 @@ namespace HondataDotNet.Datalog.OBDII
         where TReadinessTest : struct, Enum
         where TReadinessCode : IReadinessCode<TReadinessTest>
     {
-        byte FuelStatus { get; }
+        int FuelStatus { get; }
         IReadOnlyDictionary<TReadinessTest, TReadinessCode> ReadinessCodes { get; }
     }
 }
